rename publishing -> mirroring
